requirements.txt
rich
pyyaml
requests
google-generativeai
PyGithub
hypothesis
bandit
tbump
pip-audit
structlog
opentelemetry-distro
opentelemetry-exporter-otlp